// frontend/src/api/openapi/models/WhiteboardEdgeData_Input.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { WhiteboardEdgeLabelData } from "./WhiteboardEdgeLabelData";
import type { WhiteboardEdgeType } from "./WhiteboardEdgeType";
export type WhiteboardEdgeData_Input = {
  /**
   * Type of the edge
   */
  type: WhiteboardEdgeType;
  label: WhiteboardEdgeLabelData;
};